+++ src/test/java/dev/marksman/enhancediterables/ImmutableFiniteIterableTest.java
@@ -198,6 +198,23 @@ class ImmutableFiniteIterableTest {
     }
 
     @Nested
+    @DisplayName("distinct")
+    class Distinct {
+
+        @Test
+        void empty() {
+            assertThat(immutableFiniteIterable(emptyList()).distinct(), emptyIterable());
+        }
+
+        @Test
+        void removesRepeatedElementsAndRetainsOrder() {
+            assertThat(immutableFiniteIterable(asList(1, 2, 2, 3, 3, 3, 2, 2, 1, 4)).distinct(),
+                    contains(1, 2, 3, 4));
+        }
+
+    }
+
+    @Nested
     @DisplayName("drop")
     class Drop {