+++ src/test/java/dev/marksman/enhancediterables/ImmutableNonEmptyFiniteIterableTest.java
@@ -183,6 +183,18 @@ class ImmutableNonEmptyFiniteIterableTest {
     }
 
     @Nested
+    @DisplayName("distinct")
+    class Distinct {
+
+        @Test
+        void removesRepeatedElementsAndRetainsOrder() {
+            assertThat(immutableNonEmptyFiniteIterable(1, asList(2, 2, 3, 3, 3, 2, 2, 1, 4)).distinct(),
+                    contains(1, 2, 3, 4));
+        }
+
+    }
+
+    @Nested
     @DisplayName("drop")
     class Drop {